requirements.txt
openai
instructor
pydantic
python-dotenv
qdrant-client
sentence-transformers